+++ modules/sio_client/builtin/cpp/sioemitBuiltin.cpp
@@ -6,15 +6,41 @@
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 2 of the License, or
 // (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-removegateway(modulepath(nelsonroot(), 'sio_client', 'builtin'));
+#include "sioemitBuiltin.hpp"
+#include "Error.hpp"
+#include "SioClientEmit.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::SioClientGateway::sioemitBuiltin(
+    Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    if (argIn.size() < 1 || argIn.size() > 2) {
+        Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
+    }
+    if (nLhs != 0) {
+        Error(ERROR_WRONG_NUMBERS_OUTPUT_ARGS);
+    }
+    ArrayOf param1 = argIn[0];
+    std::string name = param1.getContentAsCString();
+    std::string message;
+    if (argIn.size() == 2) {
+        ArrayOf param2 = argIn[1];
+        message = param2.getContentAsCString();
+	}
+    sioemit(name, message);
+    return retval;
+}
 //=============================================================================