+++ modules/sio_client/src/cpp/SioClientEmit.cpp
@@ -6,15 +6,26 @@
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 2 of the License, or
 // (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-removegateway(modulepath(nelsonroot(), 'sio_client', 'builtin'));
+#include "SioClientEmit.hpp"
+#include "SioClientCommand.hpp"
+//=============================================================================
+namespace Nelson {
+//=============================================================================
+void
+sioemit(const std::string& name, const std::string& message)
+{
+    SioClientCommand::getInstance()->sioemit(name, message);
+}
+//=============================================================================
+}
 //=============================================================================